requirements.txt
asgiref==3.6.0
coverage==7.2.0
Django==4.1.7
django-admin-sortable2==2.1.4
django-filter==22.1
django-jalali-date==1.0.1
django-json-widget==1.1.1
django-quill-editor==0.1.40
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
future==0.18.3
jdatetime==4.1.0
Markdown==3.4.1
Pillow==9.4.0
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
upgrade-requirements==1.7.0
